+++ test/e2e/suites/operator/suite_test.go
@@ -107,8 +107,10 @@ var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
 var _ = ginkgo.SynchronizedAfterSuite(func() {
 	// cleanup all namespaces we created both in control plane and member clusters.
 	// It will not return error even if there is no such namespace in there that may happen in case setup failed.
-	err := cleanupTestNamespace(testNamespace, kubeClient)
-	gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+	if testNamespace != "" && kubeClient != nil {
+		err := cleanupTestNamespace(testNamespace, kubeClient)
+		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+	}
 }, func() {})
 
 // setupTestNamespace will create a namespace in control plane and all member clusters, most of cases will run against it.
@@ -125,6 +127,10 @@ func setupTestNamespace(namespace string, kubeClient kubernetes.Interface) error
 
 // cleanupTestNamespace will remove the namespace we set up before for the whole testing.
 func cleanupTestNamespace(namespace string, kubeClient kubernetes.Interface) error {
+	if namespace == "" || kubeClient == nil {
+		// Skip cleanup if namespace or client is not initialized
+		return nil
+	}
 	err := util.DeleteNamespace(kubeClient, namespace)
 	if err != nil {
 		return err
